requirements.txt
flask
requests
gunicorn
google-generativeai
openai
pydantic
python-dotenv
